refactor: move utils into folder
