+++ src/hub.ts
@@ -17,7 +17,7 @@ import { isContactInfo, isIDRegistryContent, isUserContent } from '~/types/typeg
 import { TypedEmitter } from 'tiny-typed-emitter';
 import RocksDB from '~/storage/db/rocksdb';
 import { err, ok, Result } from 'neverthrow';
-import { FarcasterError, ServerError } from '~/errors';
+import { FarcasterError, ServerError } from '~/utils/errors';
 import { SyncEngine } from '~/sync/syncEngine';
 
 export interface HubOpts {
