+++ src/storage/db/reaction.test.ts
@@ -2,8 +2,8 @@ import ReactionDB from '~/storage/db/reaction';
 import { Factories } from '~/test/factories';
 import { jestRocksDB } from '~/storage/db/jestUtils';
 import { Ed25519Signer, ReactionAdd, ReactionRemove } from '~/types';
-import { NotFoundError } from '~/errors';
-import { generateEd25519Signer } from '~/utils';
+import { NotFoundError } from '~/utils/errors';
+import { generateEd25519Signer } from '~/utils/utils';
 
 const rocks = jestRocksDB('db.reaction.test');
 const db = new ReactionDB(rocks);
