+++ src/storage/engine/engine.signer.test.ts
@@ -4,10 +4,10 @@ import { ResultAsync } from 'neverthrow';
 import { jestRocksDB } from '~/storage/db/jestUtils';
 import SignerDB from '~/storage/db/signer';
 import Engine from '~/storage/engine';
-import { BadRequestError, NotFoundError } from '~/errors';
+import { BadRequestError, NotFoundError } from '~/utils/errors';
 import { Factories } from '~/test/factories';
 import { Ed25519Signer, EthereumSigner, SignerAdd, SignerMessage, SignerRemove, IDRegistryEvent } from '~/types';
-import { generateEd25519Signer, generateEthereumSigner, hashFCObject } from '~/utils';
+import { generateEd25519Signer, generateEthereumSigner, hashFCObject } from '~/utils/utils';
 
 const testDb = jestRocksDB(`engine.signer.test`);
 const engine = new Engine(testDb);
