+++ src/storage/provider/idRegistryProvider.ts
@@ -3,8 +3,8 @@ import { BigNumber, Contract, providers, Event } from 'ethers';
 import { TypedEmitter } from 'tiny-typed-emitter';
 import { IDRegistryEvent } from '~/types';
 import { ok, err, Result } from 'neverthrow';
-import { sanitizeSigner } from '~/utils';
-import { BadRequestError, FarcasterError } from '~/errors';
+import { sanitizeSigner } from '~/utils/utils';
+import { BadRequestError, FarcasterError } from '~/utils/errors';
 
 export type IDRegistryEvents = {
   block: (blockNumber: number) => void;
