+++ src/storage/sets/reactionSet.test.ts
@@ -1,7 +1,7 @@
 import Faker from 'faker';
 import { jestRocksDB } from '~/storage/db/jestUtils';
 import ReactionDB from '~/storage/db/reaction';
-import { BadRequestError, NotFoundError } from '~/errors';
+import { BadRequestError, NotFoundError } from '~/utils/errors';
 import { Factories } from '~/test/factories';
 import ReactionSet from '~/storage/sets/reactionSet';
 import { Reaction, ReactionAdd, ReactionRemove, URI } from '~/types';
