+++ src/storage/sets/verificationSet.ts
@@ -1,10 +1,10 @@
 import { ResultAsync } from 'neverthrow';
 import RocksDB from '~/storage/db/rocksdb';
 import VerificationDB from '~/storage/db/verification';
-import { BadRequestError } from '~/errors';
+import { BadRequestError } from '~/utils/errors';
 import { Verification, VerificationEthereumAddress, VerificationRemove } from '~/types';
 import { isVerificationEthereumAddress, isVerificationRemove } from '~/types/typeguards';
-import { hashCompare } from '~/utils';
+import { hashCompare } from '~/utils/utils';
 
 /**
  * VerificationSet is a modified LWW set that stores and fetches verifications. VerificationEthereumAddress and VerificationRemove
