+++ src/urls/castUrl.ts
@@ -4,7 +4,7 @@ import { Result, err, ok } from 'neverthrow';
 import { UserId, UserIdConstructorArgs, UserIdParams, UserIdSpec } from '~/urls/userUrl';
 import { FarcasterURL } from '~/urls/baseUrl';
 import { isValidId, getParams, joinParams } from '~/urls/utils';
-import { BadRequestError, FarcasterError, ServerError } from '~/errors';
+import { BadRequestError, FarcasterError, ServerError } from '~/utils/errors';
 
 const REGEX_BLAKE2B_HASH = '0x[a-f0-9]{128}';
