+++ src/urls/chainAccountUrl.ts
@@ -1,7 +1,7 @@
 import { AccountId } from 'caip';
 import { ethers } from 'ethers';
 import { err, ok, Result } from 'neverthrow';
-import { BadRequestError, FarcasterError, ServerError } from '~/errors';
+import { BadRequestError, FarcasterError, ServerError } from '~/utils/errors';
 import { BaseChainURL, ChainURL } from '~/urls/chainUrl';
 
 /**
