+++ src/urls/index.ts
@@ -7,7 +7,7 @@ import { UserURL } from '~/urls/userUrl';
 import { URL } from '~/urls/baseUrl';
 import { Web2URL } from '~/urls/web2Url';
 import { ChainAccountURL } from '~/urls/chainAccountUrl';
-import { BadRequestError, FarcasterError } from '~/errors';
+import { BadRequestError, FarcasterError } from '~/utils/errors';
 
 export const parseUrl = (
   url: string,
