+++ src/urls/userUrl.ts
@@ -2,7 +2,7 @@ import { IdentifierSpec } from 'caip/dist/types';
 import { Result, err, ok } from 'neverthrow';
 import { isValidId, getParams, joinParams } from '~/urls/utils';
 import { FarcasterURL } from '~/urls/baseUrl';
-import { BadRequestError, FarcasterError } from '~/errors';
+import { BadRequestError, FarcasterError } from '~/utils/errors';
 
 export const UserIdSpec: IdentifierSpec = {
   name: 'userId',
